requirements.txt

langchain==0.3.0
langchain-openai==0.2.0
langchain-core==0.3.0
langchain-community==0.3.0
openai==1.54.3
langgraph

mysql-connector-python==9.1.0
pymysql==1.1.1
SQLAlchemy==2.0.35

fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0

python-dotenv==1.0.1
typing-extensions==4.12.2

python-dateutil==2.9.0
pytz==2024.2

SpeechRecognition==3.10.4

pyaudio==0.2.14

openai-whisper==20231117

sounddevice==0.5.1
soundfile==0.12.1
numpy==1.26.4  # Required for audio processing

gTTS==2.5.3

pyttsx3==2.98
pygame==2.6.1  # Alternative audio player

colorama==0.4.6  # Colored terminal output
rich==13.9.4     # Beautiful terminal formatting
click==8.1.7     # CLI framework

requests==2.32.3
httpx==0.27.2
aiofiles==24.1.0

# JSON handling
ujson==5.10.0

# Progress bars
tqdm==4.66.5
loguru==0.7.2

pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0


email-validator==2.2.0
phonenumbers==8.13.48

black==24.10.0  # Code formatter
flake8==7.1.1   # Linter
mypy==1.13.0    # Type checker

python-multipart==0.0.6
requests==2.31.0